+++ src/datastructures/vit.h
@@ -5,7 +5,7 @@
 typedef struct vitnode {
 	vertexid_t start;
 	vertexid_t end;
-	int degree;
+	long degree;
 }VitNode;
 
 class Vit {
@@ -16,19 +16,19 @@ class Vit {
 	
 	public:
 		Vit();
-		Vit(int size,vertexid_t *start,vertexid_t *end,int *degrees);
+		Vit(int size,vertexid_t *start,vertexid_t *end,long *degrees);
 		void clear();
 
 		// getter and setter
 		inline int getSize() {return size;}
 		inline int getStart(int vitId){return p[vitId].start;}
 		inline int getEnd(int vitId){return p[vitId].end;}
-		inline int getDegree(int vitId){return p[vitId].degree;}
+		inline long getDegree(int vitId){return p[vitId].degree;}
 	
-		inline void setDegree(int vitId,int numEdges) {p[vitId].degree = numEdges;}
+		inline void setDegree(int vitId,long numEdges) {p[vitId].degree = numEdges;}
 
-		void setVitValue(int vitId,vertexid_t start,vertexid_t end,int numEdges);
-		void add(vertexid_t start,vertexid_t end,int numEdges);
+		void setVitValue(int vitId,vertexid_t start,vertexid_t end,long numEdges);
+		void add(vertexid_t start,vertexid_t end,long numEdges);
 		void print();
 };
